Fixes #48
This time it should fix it, I replaced RestSharp with System.Http.Net and
it all just works. I guess RestSharp didn't have a proper model for when
it disconnected/kept the connection alive.

Further, introduced a bit of logging infrastructure pertaining to #17 but
haven't replaced all call-sites of Log.log yet. However, the added logging
should give enough of a view of what is going on and when TcpClient is
cancelled.

This will most likely affect performance, so the introduced interface has been
done as a 'deferred' log method that you pass a function to, so that the
function is not evaluated every time, but only if the log level is low enough
(or equal to that of the log line).