extracting log_format from Suave.Http.log
This makes it possible to nicely format the request with an expernal log framework.

Pertains to #49. Document the fact that suave is logging to a stream in a concurrent
manner; in the long run we could provide a logging actor that ships logs further to
log sinks or have a nicer way to compose the logging infrastructure.