adding TODO on deferral of body read
If we always read all requests fully into memory, we can get
DoS-attacked by just posting large data requests to the server and we
might want to perform a Content-Length header check before reading all
that data.

By performing the header check, we can respond with 100 Continue if it's
a valid length/something we accept. This will also work better in proxy
mode when we make more of the request contents available to the proxy,
since it would like to read piecemeal from the request input stream and
write to the upstream server.