Work in progress
I wanted to be able to discover the API through the IDE or by browsing by
module. Comments are both in Http.fs and Http.fsi and I haven't finished
with Http.fsi yet, as I still need to move comments there and add the newly
added server methods to it.

I started a small sub-project .WebMachine that will help build RESTful APIs with
resources.