Time is a global, DateTime -> DateTimeOffset.
MSDN has this to say:

These uses for DateTimeOffset values are much more common than those for
DateTime values. As a result, DateTimeOffset should be considered the default
date and time type for application development.

A DateTimeOffset value is not tied to a particular time zone, but can originate
from any of a variety of time zones. To illustrate this, the following example
lists the time zones to which a number of DateTimeOffset values (including a
local Pacific Standard Time) can belong.

And for DateTime:

Unless a particular DateTime value represents UTC, that date and time value is
often ambiguous or limited in its portability.

http://msdn.microsoft.com/en-us/library/bb384267.aspx