AssemblyInfo, Utils -> Http.fs
Moving utils functions for dealing with combinations of applicatives to
Http, because that's where they are used and then we don't have to use
an assembly attribute to auto-open Utils.fs.

I'm choosing to add all generated AssemblyVersionInfo.fs files until
further notice, as they make it easier for newbies to simply clone this
repository and compile it in Visual Studio or Xamarin Studio.