Invalidate Razor file cache when file changes
When using Suave in-process (F# Interactive or with automatic reloader), I wanted
to be able to change a Razor template and see the change immediately on the page.
This change makes this possible by tracking when the template was last changed and
invalidating it if the file changes. The only tricky thing is that we need to use
the time stamp as part of the Razor "template cache key" (otherwise `Razor.Parse`
throws).