[owin, BREAKING] all tests passing
This commit is breaking in that it changes a public interface.

Previously socket task has signature
`SocketTask of (Connection -> SocketOp<unit>)` but now it has
`SocketTask of (Connection * HttpResult -> SocketOp<unit>)` which was
changed in order to facilitate downstream-bound WebParts changing the
contents/headers in a way that the SocketTask needs to take into account
at time of evaluating the `socket{}` workflow.

This means that the HelloWorld test-case for OWIN is fully functional
and we may release a new version with OWIN support.

Enjoy!
// Henrik <henrik@haf.se>