Add pathScanCi and urlScanCi combinators
The pathScanCi and urlScanCi combinators are the case insensitive versions of the pathScan and urlScan combinators. They scan the request path against a format , regardless of casing, to extract dynamic values from the path and pass them on to a handler.

This commit adds a signature for the pathScanCi combinator to the Combinators.fsi file and the corresponding implementation to the Combinators.fs file. In accordance with the existing urlScan combinator, the urlScanCi combinator implementation is added to the Combinators.fs file but no signature is defined in the Combinators.fsi file.